戻り値も含めて使うので、UnityEventからデリゲートに変更。
